Add LICENSE
This commit adds a LICENSE file to the repository, specifying the terms under which the project's code can be used, modified, and distributed. The chosen license is the MIT License, which is a permissive open-source license that allows for reuse within proprietary software provided that all copies of the licensed software include a copy of the MIT License terms and the copyright notice.